requirements.txt
amqp==1.4.9
anyjson==0.3.3
APScheduler==3.5.3
billiard==3.3.0.23
celery==3.1.18
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
django-celery==3.2.2
django-jsonfield==1.0.1
django-suit==0.2.26
idna==2.8
IPy==0.83
kombu==3.0.37
mysqlclient==1.3.14
pytz==2018.7
requests==2.21.0
six==1.12.0
tzlocal==1.5.1
urllib3==1.24.1
vine==1.2.0
xlrd==1.2.0
xlwt==1.3.0
